+++ mopViewer/mainWindow.h
@@ -13,8 +13,8 @@ class mainWindow {
 public:
 mainWindow(void);
 ~mainWindow(void);
-void showStats(std::string fileName, int skipCount);
-void selectGame(std::string fileName, int skipCount);
+void showStats(std::string fileName, float skipCount);
+void selectGame(std::string fileName, float skipCount);
 void refreshAll(WINDOW * window1, WINDOW * window2);
 int printOutput(MopState * mopstate, int currentrow, int currentitem, WINDOW* window);
 MopFile * mopfile;
